pyproject.toml
[tool.poetry]
name = "environ_import"
version = "0.1.0"
description = "A library that allows you to import environment variables like any other module."
authors = ["NimajnebEC <nimajnebec@users.noreply.github.com>"]
readme = "README.md"
license = "MIT"
packages = [
    { include = "environ_import", from = "src" },
    { include = "environ", from = "src" },
]

[tool.poetry.dependencies]
python-dotenv = "^1.0.0"
python = "^3.8.1"


[tool.poetry.group.dev.dependencies]
flake8-length = "^0.3.1"
flake8 = "^6.1.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
